Added coin sound, weapon swapping, and Play As Guest transition. Portals no longer active while enemies are still in play.

+++ RogueLike201/Assets/Scripts/PlayerScript.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class PlayerScript : MonoBehaviour
 {
     public Rigidbody2D rb;
-    public int moveVal;
+    public float moveVal;
     public int coinCount;
     public GameObject crossHair;
     public GameObject weaponHolder;
